extracted increment wiki view count logic to its own file and added counter for revision wikis
